requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
openai==1.79.0         
httpx==0.27.2           
python-dotenv==1.0.1
chromadb==1.0.10
sentence-transformers==4.1.0